+++ utils/motors_wrapper.h
@@ -10,7 +10,7 @@ struct motors_wrapper {
 
 void motors_wrapper_init(struct ausbee_l298_chip *right_motor,
                          struct ausbee_l298_chip *left_motor);
-void motors_wrapper_right_motor_set_duty_cycle(void *dummy, int32_t duty_cycle);
-void motors_wrapper_left_motor_set_duty_cycle(void *dummy, int32_t duty_cycle);
+void motors_wrapper_right_motor_set_duty_cycle(void *dummy, float duty_cycle);
+void motors_wrapper_left_motor_set_duty_cycle(void *dummy, float duty_cycle);
 
 #endif /* MOTORS_WRAPPER_H */
